+++ include/editor.h
@@ -43,6 +43,8 @@ public:
     void redo();
     void closeProject();
     bool setMap(QString map_name);
+    void saveUiFields();
+    void saveEncounterTabData();
     void displayMap();
     void displayMetatileSelector();
     void displayMapMetatiles();
@@ -56,11 +58,13 @@ public:
     void displayMapConnections();
     void displayMapBorder();
     void displayMapGrid();
+    void displayWildMonTables();
 
     void setEditingMap();
     void setEditingCollision();
     void setEditingObjects();
     void setEditingConnections();
+    void clearWildMonTabWidgets();
     void setCurrentConnectionDirection(QString curDirection);
     void updateCurrentConnectionDirection(QString curDirection);
     void setConnectionsVisibility(bool visible);
@@ -68,6 +72,7 @@ public:
     void setConnectionMap(QString mapName);
     void addNewConnection();
     void removeCurrentConnection();
+    void addNewWildMonGroup();
     void updateDiveMap(QString mapName);
     void updateEmergeMap(QString mapName);
     void setSelectedConnectionFromMap(QString mapName);
@@ -75,6 +80,7 @@ public:
     void updateSecondaryTileset(QString tilesetLabel, bool forceLoad = false);
     void toggleBorderVisibility(bool visible);
     void updateCustomMapHeaderValues(QTableWidget *);
+    void configureEncounterJSON();
     Tileset *getCurrentMapPrimaryTileset();
 
     DraggablePixmapItem *addMapEvent(Event *event);